opus-trencher-bot-v3/opus-trencher-bot/requirements.txt
python-telegram-bot
python-dotenv
aiohttp
websockets
solders
solana
base58
requests
numpy
pandas
